back-end/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14,<4.0
psycopg2-binary>=2.9,<3.0
django-environ>=0.9.0,<1.0
